Aragon/vip.py
#!/usr/bin/env python3
# -*- coding: utf-8 -*-

import logging
from telegram.ext import CommandHandler

logger = logging.getLogger(__name__)

async def vip_command(update, context):
    await update.message.reply_text("VIP system coming soon...")

async def check_vip_status(user_id):
    return None

def setup_vip_handlers(application):
    application.add_handler(CommandHandler("vip", vip_command))
